+++ wayland.c
@@ -849,7 +849,7 @@ handle_global(void *data, struct wl_registry *registry,
     }
 
     else if (strcmp(interface, wl_data_device_manager_interface.name) == 0) {
-        const uint32_t required = 1;
+        const uint32_t required = 3;
         if (!verify_iface_version(interface, version, required))
             return;